requirements.txt
django_heroku
requests
geopy
Django
beautifulsoup4
gunicorn
html5lib
joblib
pandas
sklearn
celery
rabbitmq
honcho
eventlet
opencv-python-headless
matplotlib
keras
tensorflow-cpu==2.5.0
PyGithub